// multisto/src/Schema.ts
export interface Currency {
  ID: number;
  Country?: string | null;
  Currency?: string | null;
  Abbreviation?: string | null;
  Symbol?: string | null;
  /**  Defaults to: 0. */
  isBlockchainBased?: number | null;
  /**  Defaults to: 1. */
  blockchainID: number;
  /**  Defaults to: 0. */
  isNative: number;
  /**  Defaults to: 0. */
  Address: string;
  /**  Defaults to: 0. */
  cryptoReceivingAddress: string;
}
// ...
export interface Investor {
  ID: number;
  FirstName?: string | null;
  LastName?: string | null;
  Address?: string | null;
  country?: string | null;
  phone?: string | null;
  cell?: string | null;
  zip?: string | null;
  town?: string | null;
  state?: string | null;
  Password: string;
  email: string;
  PassportNumber?: string | null;
  NationalID?: string | null;
  /**  Defaults to: 0. */
  receiveEmails?: number | null;
  kinname?: string | null;
  kinphone?: string | null;
  kinemail?: string | null;
  /**  Defaults to: 0. */
  investorType: number;
  CompanyName?: string | null;
  TitleWithinCompany?: string | null;
  /**  Defaults to: 0. */
  PowerToBindCompany?: number | null;
  DOB?: Date | null;
  /**  Defaults to: 0. */
  twofactorenable?: number | null;
  /**  Defaults to: 0. */
  investorBulkUploadStatus?: number | null;
  /**  Defaults to: en. */
  language?: string | null;
  MiddleName?: string | null;
  SocialSecurity?: string | null;
  MailingAddress?: string | null;
  FaxNumber?: string | null;
  /**  Defaults to: 0. */
  MaritalStatus?: number | null;
  Occupation?: string | null;
  EmployerName?: string | null;
  EmployerAddress?: string | null;
  /**  Defaults to: 0. */
  RetirementAccount?: number | null;
  /**  Defaults to: 0. */
  TrustOrBusinessEntity?: number | null;
  DateIncorporation?: Date | null;
  TaxIDNo?: string | null;
  GovtIDNo?: string | null;
  /**  Defaults to: 0. */
  GovtIDNoIsTaxNo?: number | null;
  NameSecondaryContact?: string | null;
  PhoneSecondaryContact?: string | null;
  EmailSecondaryContact?: string | null;
  NamePrimaryContact?: string | null;
  PhonePrimaryContact?: string | null;
  EmailPrimaryContact?: string | null;
  PrincipalCountryOfBusiness?: string | null;
  countryOfCitizenship?: string | null;
  /**  Defaults to: 0. */
  referByInvestorID?: number | null;
  kyc?: string | null;
  /**  Defaults to: 1. */
  dividendPeriod?: number | null;
  DOBCountry?: string | null;
  taxResidencyCountry?: string | null;
  /**  Defaults to: 0. */
  affiliateStatus: number;
  /**  Defaults to: 0. */
  PoliticallyExposedPerson: number;
  PoliticallyExposedPersonPost?: string | null;
  mercuryID?: string | null;
  uploadDate?: Date | null;
  twoFactorCode?: number | null;
  brokerID?: string | null;
  driversLicenseID?: string | null;
  /**  Defaults to: CURRENT_TIMESTAMP. */
  investmentLimitUpdateDate: Date;
  /**  Defaults to: 0. */
  allowedTotalInvestment: number;
  /**  Defaults to: 0. */
  yearlyTotalInvestment: number;
}
// ...
export interface Shares {
  ID: number;
  /**  Defaults to: 0. */
  stoid?: number | null;
  shareTypeid: number;
  /**  Defaults to: []. */
  PublicKey?: string | null;
  /**  Defaults to: 0. */
  isBlockchainFrozen: number;
  /**  Defaults to: 0. */
  isBlockchainAuthorized: number;
  /**  Defaults to: 0.000. */
  shares?: number | null;
  investorID: number;
  sharesHistoryID: number;
}
// ...
export interface Sharetypes {
  ID: number;
  title: string;
  stoid: number;
  /**  Defaults to: 0.000. */
  totalShares?: number | null;
  /**  Defaults to: 0.000. */
  companyShares?: number | null;
  /**  Defaults to: 0.0000000000000000. */
  nominalValue?: number | null;
  /**  Defaults to: 0. */
  isNominalValueApplicable: number;
  /**  Defaults to: 0. */
  isVotingRightsApplicable: number;
  /**  Defaults to: 0. */
  isDividendRightsApplicable: number;
  /**  Defaults to: 0. */
  isblockchain: number;
  ethereumContractAddress: string;
  ethereumWhitelistAddress: string;
  /**  Defaults to: 0.0000000000000000. */
  premimum?: number | null;
  currencyid: number;
  /**  Defaults to: 1. */
  needauthorization: number;
  token_abi?: string | null;
  whitelist_abi?: string | null;
  ethereumBlockchainPublicAddress?: string | null;
  /**  Defaults to: default. */
  subscriptionform?: string | null;
  /**  Defaults to: 0.0000. */
  minimumSharesToBuyByInvestor?: number | null;
  /**  Defaults to: 0. */
  blockchainProtocol?: number | null;
  /**  Defaults to: 0. */
  blockchainBuyOrdersAllowed?: number | null;
  /**  Defaults to: 0.000. */
  reduceSharesForPurchase?: number | null;
  /**  Defaults to: 1. */
  isEnabled?: number | null;
  /**  Defaults to: 0. */
  walletCustodayType?: number | null;
  tanganyWalletID?: string | null;
  /**  Defaults to: 0. */
  investorCanPurchaseDirectly?: number | null;
  AssetName?: string | null;
  AssetTag?: string | null;
  /**  Defaults to: 1.00. */
  votingPower: number;
  /**  Defaults to: 1. */
  isMeetingRightsApplicable?: number | null;
  /**  Defaults to: 1. */
  isInvestorTradable?: number | null;
  /**  Defaults to: 18. */
  blockchainDecimals?: number | null;
  ipfsDocumentHash?: string | null;
  /**  Defaults to: 0.000. */
  custodianShares: number;
  /**  Defaults to: 1. */
  sellToCompany: number;
  /**  Defaults to: 0.0000000000000000. */
  sellValue: number;
  /**  Defaults to: 0. */
  isShareNosApplicable: number;
  /**  Defaults to: 0. */
  isCertificateNosApplicable: number;
  /**  Defaults to: -1. */
  channelIDForAutoPayments?: number | null;
  keyStoreFileAutoPayments?: string | null;
  keyStoreFileAutoPaymentsPassword?: string | null;
}